+++ api/rust/cargo/lief/src/generic.rs
@@ -117,6 +117,12 @@ impl std::fmt::Debug for &dyn Relocation {
     }
 }
 
+/// Generic interface representing a binary executable.
+///
+/// This trait provides a unified interface across multiple binary formats
+/// such as ELF, PE, Mach-O, and others. It enables users to access binary
+/// components like headers, sections, symbols, relocations,
+/// and functions in a format-agnostic way.
 pub trait Binary {
     #[doc(hidden)]
     fn as_generic(&self) -> &ffi::AbstractBinary;
@@ -167,7 +173,7 @@ pub trait Binary {
         into_optional(self.as_generic().debug_info())
     }
 
-    /// Disassemble code starting a the given virtual address and with the given
+    /// Disassemble code starting at the given virtual address and with the given
     /// size.
     ///
     /// ```
@@ -248,6 +254,26 @@ pub trait Binary {
     /// current binary
     fn page_size(&self) -> u64 {
         self.as_generic().page_size()
+    }
+
+    /// Load and associate an external debug file (e.g., DWARF or PDB) with this binary.
+    ///
+    /// This method attempts to load the debug information from the file located at the given path,
+    /// and binds it to the current binary instance. If successful, it returns the
+    /// loaded [`crate::DebugInfo`] object.
+    ///
+    /// <div class="warning">
+    /// It is the caller's responsibility to ensure that the debug file is
+    /// compatible with the binary. Incorrect associations may lead to
+    /// inconsistent or invalid results.
+    /// </div>
+    ///
+    /// <div class="note">
+    /// This function does not verify that the debug file matches the binary's unique
+    /// identifier (e.g., build ID, GUID).
+    /// </div>
+    fn load_debug_info(&mut self, path: &std::path::Path) -> Option<crate::DebugInfo> {
+        into_optional(self.as_pin_mut_generic().load_debug_info(path.to_str().unwrap()))
     }
 }