+++ src/Abstract/Binary.cpp
@@ -70,10 +70,10 @@ std::vector<uint64_t> Binary::xref(uint64_t address) const {
   return result;
 }
 
-
 uint64_t Binary::page_size() const {
   return get_pagesize(*this);
 }
+
 
 void Binary::accept(Visitor& visitor) const {
   visitor.visit(*this);